fix: improve image selector logic
